new Score controller with drawing method
