+++ app/javascript/controllers/score_controller.js
@@ -7,9 +7,10 @@
 //   <h1 data-target="hello.output"></h1>
 // </div>
 
-import Vex from 'vexflow'
+
 import { Controller } from "stimulus"
 import { Music } from "../models/music"
+import { Score } from "../models/score"
 
 export default class extends Controller {
   static targets = [ "output" ]
@@ -19,12 +20,12 @@ export default class extends Controller {
 
   connect() {
     this.music = new Music(this.notesValue, 0)
+    this.score = new Score(this.music)
     this.initConverters();
     this.currentSelection = null;
-    const VF = Vex.Flow;
-    this.vf = new Vex.Flow.Factory({renderer: {elementId: 'score'}})
 
-    this.draw();
+
+    this.score.draw();
     this.updateAttemptStringPlayback();
   }
 
@@ -47,36 +48,6 @@ export default class extends Controller {
       this.noteName2MidiNum[noteNameSharp + String(octave)] = midiNumShift+i
       this.noteName2MidiNum[noteNameFlat + String(octave)] = midiNumShift+i // if the noteNameFlat==noteNameSharp, it's just overwritten
     }
-  }
-
-  draw(event) {
-    if (event) {
-      event.preventDefault();
-    }
-    this.vf.context.clear();
-    const score = this.vf.EasyScore();
-    const system = this.vf.System();
-
-    system.addStave({
-      voices: [score.voice(score.notes(this.music.getVexString()))]
-    }).addClef('treble').addTimeSignature('4/4');
-
-    this.vf.draw();
-
-    this.addActionToNotes()
-    this.addTabIndexToNotes();
-  }
-
-  addActionToNotes() {
-    const svg = document.querySelector("svg");
-    const notes = svg.querySelectorAll(".vf-stavenote");
-    notes.forEach((note) => {note.setAttribute("data-action", "click->score#clickNote")});
-  }
-
-  addTabIndexToNotes() {
-    const svg = document.querySelector("svg");
-    const notes = svg.querySelectorAll(".vf-stavenote");
-    notes.forEach((note) => {note.setAttribute("tabindex", "0")});
   }
 
   clickNote(event) {
@@ -170,7 +141,7 @@ export default class extends Controller {
       }
     }
 
-    this.draw(event);
+    this.score.draw(event);
     this.updateAttemptStringPlayback(event);
     const svgNote = this.getSvgNoteFromIndex(index);
     this.toggleNoteSelection(svgNote);
